adaptive: Update module names
