+++ src/core/adaptive/adaptive_representation_selector.ts
@@ -26,10 +26,12 @@ import createSharedReference, {
   IReadOnlySharedReference,
   ISharedReference,
 } from "../../utils/reference";
+import takeFirstSet from "../../utils/take_first_set";
 import TaskCanceller, {
   CancellationSignal,
 } from "../../utils/task_canceller";
 import { IReadOnlyPlaybackObserver } from "../api";
+import { IBufferType } from "../segment_buffers";
 import BufferBasedChooser from "./buffer_based_chooser";
 import GuessBasedChooser from "./guess_based_chooser";
 import NetworkAnalyzer from "./network_analyzer";
@@ -47,6 +49,99 @@ import RepresentationScoreCalculator from "./utils/representation_score_calculat
 import selectOptimalRepresentation from "./utils/select_optimal_representation";
 
 /**
+ * Select the most adapted Representation according to the network and buffer
+ * metrics it receives.
+ * @param {Object} options
+ */
+export default function createAdaptiveRepresentationSelector(
+  options : IAdaptiveRepresentationSelectorArguments
+) : IRepresentationEstimator {
+  /**
+   * Allows to estimate the current network bandwidth.
+   * One per active media type.
+   */
+  const bandwidthEstimators : Partial<Record<IBufferType, BandwidthEstimator>> = {};
+  const manualBitrates = options.manualBitrates;
+  const minAutoBitrates = options.minAutoBitrates;
+  const maxAutoBitrates = options.maxAutoBitrates;
+  const initialBitrates = options.initialBitrates;
+  const throttlers = options.throttlers;
+  const lowLatencyMode = options.lowLatencyMode;
+
+  /**
+   * Take type and an array of the available representations, spit out an
+   * observable emitting the best representation (given the network/buffer
+   * state).
+   * @param {string} type
+   * @param {Object} currentRepresentation - Reference emitting the
+   * Representation currently loaded.
+   * @param {Object} representations - Reference emitting the list of available
+   * Representations to choose from.
+   * @param {Object} playbackObserver - Emits regularly playback conditions
+   * @param {Object} cancellationSignal - After this `CancellationSignal` emits,
+   * estimates will stop to be emitted.
+   * @returns {Array.<Object>}
+   */
+  return function getEstimates(
+    type : IBufferType,
+    currentRepresentation : IReadOnlySharedReference<Representation | null>,
+    representations : IReadOnlySharedReference<Representation[]>,
+    playbackObserver : IReadOnlyPlaybackObserver<
+      IRepresentationEstimatorPlaybackObservation
+    >,
+    cancellationSignal : CancellationSignal
+  ) : [ IReadOnlySharedReference<IABREstimate>,
+        IRepresentationEstimatorCallbacks ] {
+    const bandwidthEstimator = _getBandwidthEstimator(type);
+    const manualBitrate = takeFirstSet<IReadOnlySharedReference<number>>(
+      manualBitrates[type],
+      createSharedReference(-1));
+    const minAutoBitrate = takeFirstSet<IReadOnlySharedReference<number>>(
+      minAutoBitrates[type],
+      createSharedReference(0));
+    const maxAutoBitrate = takeFirstSet<IReadOnlySharedReference<number>>(
+      maxAutoBitrates[type],
+      createSharedReference(Infinity));
+    const initialBitrate = takeFirstSet<number>(initialBitrates[type], 0);
+    const filters = {
+      limitWidth: takeFirstSet<IReadOnlySharedReference<number | undefined>>(
+        throttlers.limitWidth[type],
+        createSharedReference(undefined)),
+      throttleBitrate: takeFirstSet<IReadOnlySharedReference<number>>(
+        throttlers.throttleBitrate[type],
+        throttlers.throttle[type],
+        createSharedReference(Infinity)),
+    };
+    return getEstimateReference({ bandwidthEstimator,
+                                  currentRepresentation,
+                                  filters,
+                                  initialBitrate,
+                                  manualBitrate,
+                                  minAutoBitrate,
+                                  maxAutoBitrate,
+                                  playbackObserver,
+                                  representations,
+                                  lowLatencyMode },
+                                cancellationSignal);
+  };
+
+  /**
+   * @param {string} bufferType
+   * @returns {Object}
+   */
+  function _getBandwidthEstimator(bufferType : IBufferType) : BandwidthEstimator {
+    const originalBandwidthEstimator = bandwidthEstimators[bufferType];
+    if (originalBandwidthEstimator == null) {
+      log.debug("ABR: Creating new BandwidthEstimator for ", bufferType);
+      const bandwidthEstimator = new BandwidthEstimator();
+      bandwidthEstimators[bufferType] = bandwidthEstimator;
+      return bandwidthEstimator;
+    }
+    return originalBandwidthEstimator;
+  }
+}
+
+/**
  * Estimate regularly the current network bandwidth and the best Representation
  * that can be played according to the current network and playback conditions.
  *
@@ -59,7 +154,7 @@ import selectOptimalRepresentation from "./utils/select_optimal_representation";
  * @param {Object} cancellationSignal
  * @returns {Array.<Object>}
  */
-export default function getEstimateReference(
+function getEstimateReference(
   { bandwidthEstimator,
     currentRepresentation,
     filters,
@@ -671,4 +766,56 @@ export interface IRepresentationEstimatorArguments {
   maxAutoBitrate : IReadOnlySharedReference<number>;
   /** The list of Representations `getEstimateReference` can choose from. */
   representations : IReadOnlySharedReference<Representation[]>;
+}
+
+/**
+ * Type of the function returned by `createAdaptiveRepresentationSelector`,
+ * allowing to estimate the most adapted `Representation`.
+ */
+export type IRepresentationEstimator = (
+  type : IBufferType,
+  currentRepresentation : IReadOnlySharedReference<Representation | null>,
+  representations : IReadOnlySharedReference<Representation[]>,
+  playbackObserver : IReadOnlyPlaybackObserver<
+    IRepresentationEstimatorPlaybackObservation
+  >,
+  cancellationSignal : CancellationSignal
+) => [ IReadOnlySharedReference<IABREstimate>,
+       IRepresentationEstimatorCallbacks ];
+
+/** Arguments received by `createAdaptiveRepresentationSelector`. */
+export interface IAdaptiveRepresentationSelectorArguments {
+  /** Initial bitrate chosen, per type (minimum if not set) */
+  initialBitrates: Partial<Record<IBufferType, number>>;
+
+  /**
+   * Some settings can depend on wether you're playing a low-latency content.
+   * Set it to `true` if you're playing such content.
+   */
+  lowLatencyMode: boolean;
+
+  /** Minimum bitrate chosen when in auto mode, per type (0 by default) */
+  minAutoBitrates: Partial<Record<IBufferType, IReadOnlySharedReference<number>>>;
+
+  /** Maximum bitrate chosen when in auto mode, per type (0 by default) */
+  maxAutoBitrates: Partial<Record<IBufferType, IReadOnlySharedReference<number>>>;
+
+  /** Manually forced bitrate set for a given type (`-1` for auto mode */
+  manualBitrates: Partial<Record<IBufferType, IReadOnlySharedReference<number>>>;
+
+  /** Allows to filter which Representations can be choosen. */
+  throttlers: IRepresentationEstimatorThrottlers;
+}
+
+/**
+ * Throttlers are interfaces allowing to limit the pool of Representations
+ * to choose from.
+ */
+export interface IRepresentationEstimatorThrottlers {
+  limitWidth : Partial<Record<IBufferType,
+                              IReadOnlySharedReference<number>>>;
+  throttle : Partial<Record<IBufferType,
+                            IReadOnlySharedReference<number>>>;
+  throttleBitrate : Partial<Record<IBufferType,
+                                   IReadOnlySharedReference<number>>>;
 }
